getting rid of old, commented code

+++ Classes/RZDebugMenuSettingsInterface.m
@@ -147,7 +147,7 @@ static NSString * const kRZVersionInfoReuseIdentifier = @"version";
 
 - (RZDebugMenuSettingsItem *)settingsItemAtIndexPath:(NSIndexPath *)indexPath
 {
-    int numberOfItems = self.settingsCellItemsMetaData.count;
+    NSUInteger numberOfItems = self.settingsCellItemsMetaData.count;
     if ( indexPath.section > 0 || indexPath.row >= numberOfItems ) {
         return nil;
     }
